requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
twilio==8.5.0
discord-webhook==1.3.0
schedule==1.2.0